refactor(app): split init logic in separate Application module
main.cpp now only creates an Application object and calls run.
This way main stays clean while the Application objects deals with init.
