+++ src/psm_gui/CMakeLists.txt
@@ -1,17 +1,6 @@
-add_executable(psm_gui main.cpp)
-
 find_package(glfw3 REQUIRED)
 find_package(OpenGL REQUIRED)
 find_package(imgui REQUIRED)
 
-target_include_directories(psm_gui PRIVATE ${CMAKE_CURRENT_SOURCE_DIR})
-
-target_link_libraries(psm_gui PRIVATE glfw OpenGL::GL imgui::imgui)
-
-if(MSVC)
-  add_custom_command(
-    TARGET psm_gui
-    POST_BUILD
-    COMMAND ${CMAKE_COMMAND} -E copy -t $<TARGET_FILE_DIR:psm_gui>
-    COMMAND_EXPAND_LISTS)
-endif()
+add_subdirectory(app)
+add_subdirectory(ui)
