+++ src/psm_gui/app/Application.hpp
@@ -1,0 +1,19 @@
+#pragma once
+
+#include <GLFW/glfw3.h>
+
+namespace psm_gui {
+class Application {
+ public:
+  Application();
+  ~Application();
+  int run();
+
+ private:
+  GLFWwindow* window_;
+  static void error_callback([[maybe_unused]] int error, const char* desc);
+  static void framebuffer_size_callback([[maybe_unused]] GLFWwindow* window,
+                                        int width, int height);
+};
+
+}  // namespace psm_gui
