--- src/psm_gui/main.cpp
@@ -1,63 +1,0 @@
-#include <GLFW/glfw3.h>
-#include <imgui.h>
-#include <imgui_impl_glfw.h>
-#include <imgui_impl_opengl3.h>
-
-#include <iostream>
-
-void glfw_error_callback([[maybe_unused]] int error, const char* desc) {
-  std::cout << "Error: " << desc << "\n";
-}
-
-void framebuffer_size_callback([[maybe_unused]] GLFWwindow* window, int width,
-                               int height) {
-  glViewport(0, 0, width, height);
-}
-
-int main() {
-  // GL 3.0 + GLSL 130
-  const char* glsl_version = "#version 130";
-  glfwInit();
-  glfwWindowHint(GLFW_CONTEXT_VERSION_MAJOR, 3);
-  glfwWindowHint(GLFW_CONTEXT_VERSION_MINOR, 3);
-  glfwSetErrorCallback(glfw_error_callback);
-
-  GLFWwindow* window =
-      glfwCreateWindow(800, 600, "PSM Gui Demo", nullptr, nullptr);
-  if (window == nullptr) {
-    std::cout << "Failed to create GLFW window " << "\n";
-    glfwTerminate();
-    return -1;
-  }
-  glfwMakeContextCurrent(window);
-  glfwSetFramebufferSizeCallback(window, framebuffer_size_callback);
-
-  IMGUI_CHECKVERSION();
-  ImGui::CreateContext();
-  ImGuiIO& io = ImGui::GetIO();
-  io.ConfigFlags |= ImGuiConfigFlags_NavEnableKeyboard;
-  io.IniFilename = nullptr;
-
-  ImGui_ImplGlfw_InitForOpenGL(window, true);
-  ImGui_ImplOpenGL3_Init(glsl_version);
-
-  while (!glfwWindowShouldClose(window)) {
-    glfwPollEvents();
-    glfwSwapBuffers(window);
-
-    ImGui_ImplOpenGL3_NewFrame();
-    ImGui_ImplGlfw_NewFrame();
-    ImGui::NewFrame();
-    ImGui::ShowDemoWindow();
-    ImGui::Render();
-    glClear(GL_COLOR_BUFFER_BIT);
-    ImGui_ImplOpenGL3_RenderDrawData(ImGui::GetDrawData());
-  }
-
-  ImGui_ImplOpenGL3_Shutdown();
-  ImGui_ImplGlfw_Shutdown();
-  ImGui::DestroyContext();
-
-  glfwDestroyWindow(window);
-  return 0;
-}
